+++ CHANGELOG.md
@@ -10,9 +10,13 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
-- [Remove `futures` feature. Use `futures03` feature instead.](https://github.com/taiki-e/auto_enums/pull/124)
+- Remove `futures` feature. Use `futures03` feature instead. ([#124](https://github.com/taiki-e/auto_enums/pull/124))
 
-- [Merge `auto_enums_core` and `auto_enums_derive` crates into main `auto_enums` crate.](https://github.com/taiki-e/auto_enums/pull/123)
+- Merge `auto_enums_core` and `auto_enums_derive` crates into main `auto_enums` crate. ([#123](https://github.com/taiki-e/auto_enums/pull/123))
+
+- Forbid custom `Unpin`/`Drop` impl if trait has `Pin<&mut self>` receiver. See [#135](https://github.com/taiki-e/auto_enums/pull/123) for more.
+
+- Update `derive_utils` to 0.12. ([#135](https://github.com/taiki-e/auto_enums/pull/123))
 
 ## [0.7.12] - 2021-01-05